update user
update user

+++ Modules/AvnUser/Routes/web.php
@@ -14,10 +14,18 @@
 Route::prefix('user')->group(function() {
 
     // Chuyên gia
-    Route::get('/list-partern', 'ParternController@listPartern')->name('list-partern');
-    Route::get('/add-partern', 'ParternController@addPartern')->name('add-partern');
-    Route::get('/edit-partern/{id}', 'ParternController@editPartern')->name('edit-partern');
-    Route::post('/store-partern', 'ParternController@storePartern')->name('store-partern');
-    Route::put('/update-partern/{id}', 'ParternController@updatePartern')->name('update-partern');
-    Route::delete('/delete-partern/{id}', 'ParternController@deletePartern')->name('delete-partern');
+    Route::get('/list-partern', 'ParternController@listPartern')->middleware(['auth', 'permission'])->name('list-partern');
+    Route::get('/add-partern', 'ParternController@addPartern')->middleware(['auth', 'permission'])->name('add-partern');
+    Route::get('/edit-partern/{id}', 'ParternController@editPartern')->middleware(['auth', 'permission'])->name('edit-partern');
+    Route::post('/store-partern', 'ParternController@storePartern')->middleware(['auth', 'permission'])->name('store-partern');
+    Route::put('/update-partern/{id}', 'ParternController@updatePartern')->middleware(['auth', 'permission'])->name('update-partern');
+    Route::delete('/delete-partern/{id}', 'ParternController@deletePartern')->middleware(['auth', 'permission'])->name('delete-partern');
+
+    // Khách
+    Route::get('/list-customer', 'CustomerController@listCustomer')->middleware(['auth', 'permission'])->name('list-customer');
+    Route::get('/add-customer', 'CustomerController@addCustomer')->middleware(['auth', 'permission'])->name('add-customer');
+    Route::get('/edit-customer/{id}', 'CustomerController@editCustomer')->middleware(['auth', 'permission'])->name('edit-customer');
+    Route::post('/store-customer', 'CustomerController@storeCustomer')->middleware(['auth', 'permission'])->name('store-customer');
+    Route::put('/update-customer/{id}', 'CustomerController@updateCustomer')->middleware(['auth', 'permission'])->name('update-customer');
+    Route::delete('/delete-customer/{id}', 'CustomerController@deleteCustomer')->middleware(['auth', 'permission'])->name('delete-customer');
 });
